# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: OpenCloudTouch
  description: Open-Source replacement for discontinued streaming device cloud features
  version: dev-6b24c3b
paths:
...
  /api/setup/wizard/check-ports:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Check Ports
      description: Check if SSH port is accessible (Wizard Step 3).
      operationId: wizard_check_ports_api_setup_wizard_check_ports_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortCheckRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortCheckResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/backup:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Backup
      description: Create complete backup to USB stick (Wizard Step 4).
      operationId: wizard_backup_api_setup_wizard_backup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/modify-config:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Modify Config
      description: Modify OverrideSdkPrivateCfg.xml (Wizard Step 5).
      operationId: wizard_modify_config_api_setup_wizard_modify_config_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigModifyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigModifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/modify-hosts:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Modify Hosts
      description: Modify /etc/hosts (Wizard Step 6).
      operationId: wizard_modify_hosts_api_setup_wizard_modify_hosts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostsModifyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostsModifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/restore-config:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Restore Config
      description: Restore config from backup (Wizard Step 8).
      operationId: wizard_restore_config_api_setup_wizard_restore_config_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/restore-hosts:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Restore Hosts
      description: Restore hosts from backup (Wizard Step 8).
      operationId: wizard_restore_hosts_api_setup_wizard_restore_hosts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/list-backups:
    post:
      tags:
      - Setup Wizard
      summary: Wizard List Backups
      description: List available backups (Wizard Step 8).
      operationId: wizard_list_backups_api_setup_wizard_list_backups_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListBackupsRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBackupsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/reboot-device:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Reboot Device
      description: Reboot SoundTouch device via SSH (Wizard Step 7).
      operationId: wizard_reboot_device_api_setup_wizard_reboot_device_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectivityCheckRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Wizard Reboot Device Api Setup Wizard Reboot Device
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/account-pairing:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Account Pairing
      description: 'Ensure device has a margeAccountUUID (Wizard Step - Account Pairing).


        Checks if the device already has a UUID. If not, generates one and

        sets it via Telnet. Persists the UUID in the device repository for

        streaming endpoint resolution.'
      operationId: wizard_account_pairing_api_setup_wizard_account_pairing_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountPairingRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPairingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/ensure-account:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Ensure Account
      description: 'Ensure device has a margeAccountUUID (Wizard Step � after config/hosts).


        Devices without a margeAccountUUID cannot play presets (INVALID_SOURCE).

        This endpoint checks GET :8090/info and sets a UUID via Telnet if missing.


        Safe to call multiple times � no-op if UUID already present.'
      operationId: wizard_ensure_account_api_setup_wizard_ensure_account_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountPairingRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPairingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/init-persistence:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Init Persistence
      description: 'Initialize persistence files on factory-reset devices (Wizard
        Step — after account pairing).


        Factory-reset devices lack SystemConfigurationDB.xml and Sources.xml.

        Without them, the firmware never fully initialises playback state,

        causing INVALID_SOURCE on preset recall (GitHub Issue #167).


        Only creates files that are missing — never overwrites existing ones.

        Safe to call multiple times.'
      operationId: wizard_init_persistence_api_setup_wizard_init_persistence_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitPersistenceRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitPersistenceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/complete:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Complete
      description: Mark wizard setup as complete for a device.
      operationId: wizard_complete_api_setup_wizard_complete_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WizardCompleteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WizardCompleteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/setup/wizard/verify-redirect:
    post:
      tags:
      - Setup Wizard
      summary: Wizard Verify Redirect
      description: Verify a domain is redirected to OCT on the device (Wizard Step
        7).
      operationId: wizard_verify_redirect_api_setup_wizard_verify_redirect_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRedirectRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyRedirectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AccountPairingRequest:
      properties:
        device_ip:
          type: string
          title: Device Ip
          description: Device IP address
        device_id:
          type: string
          title: Device Id
          description: Device ID (MAC address)
      type: object
      required:
      - device_ip
      - device_id
      title: AccountPairingRequest
      description: Request to ensure device has a margeAccountUUID.
    AccountPairingResponse:
      properties:
        success:
          type: boolean
          title: Success
        had_uuid:
          type: boolean
          title: Had Uuid
          description: True if UUID was already present
          default: false
        uuid:
          type: string
          title: Uuid
          description: The current or newly set UUID
          default: ''
        message:
          type: string
          title: Message
          default: ''
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - success
      title: AccountPairingResponse
      description: Response from account pairing.
...
    BackupRequest:
      properties:
        device_ip:
          type: string
          title: Device Ip
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
          description: Device identifier for unique backup filenames
      type: object
      required:
      - device_ip
      title: BackupRequest
      description: Request to create device backup.
    BackupResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
        volumes:
          items:
            type: object
          type: array
          title: Volumes
        total_size_mb:
          type: number
          title: Total Size Mb
          default: 0.0
        total_duration_seconds:
          type: number
          title: Total Duration Seconds
          default: 0.0
      type: object
      required:
      - success
      - message
      title: BackupResponse
      description: Response with backup results.
...
    ConfigModifyRequest:
      properties:
        device_ip:
          type: string
          title: Device Ip
        target_addr:
          type: string
          title: Target Addr
          description: OCT server URL (e.g., http://192.168.1.100:7777 or oct.local)
      type: object
      required:
      - device_ip
      - target_addr
      title: ConfigModifyRequest
      description: Request to modify config file.
    ConfigModifyResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
        backup_path:
          type: string
          title: Backup Path
          default: ''
        diff:
          type: string
          title: Diff
          default: ''
        old_url:
          type: string
          title: Old Url
          default: ''
        new_url:
          type: string
          title: New Url
          default: ''
      type: object
      required:
      - success
      - message
      title: ConfigModifyResponse
      description: Response with config modification result.
...
    ConnectivityCheckRequest:
      properties:
        ip:
          type: string
          title: Ip
      type: object
      required:
      - ip
      title: ConnectivityCheckRequest
      description: Request to check device connectivity.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HostsModifyRequest:
      properties:
        device_ip:
          type: string
          title: Device Ip
        target_addr:
          type: string
          title: Target Addr
          description: OCT server URL (e.g., http://192.168.1.100:7777)
        include_optional:
          type: boolean
          title: Include Optional
          default: true
      type: object
      required:
      - device_ip
      - target_addr
      title: HostsModifyRequest
      description: Request to modify hosts file.
    HostsModifyResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
        backup_path:
          type: string
          title: Backup Path
          default: ''
        diff:
          type: string
          title: Diff
          default: ''
      type: object
      required:
      - success
      - message
      title: HostsModifyResponse
      description: Response with hosts modification result.
    InitPersistenceRequest:
      properties:
        device_ip:
          type: string
          title: Device Ip
        device_name:
          type: string
          maxLength: 100
          title: Device Name
          description: Device name from GET :8090/info <name>
        account_uuid:
          type: string
          pattern: ^\d{1,10}$
          title: Account Uuid
          description: margeAccountUUID (7-digit numeric, from account pairing)
      type: object
      required:
      - device_ip
      - device_name
      - account_uuid
      title: InitPersistenceRequest
      description: Request to initialize persistence files on a factory-reset device.
    InitPersistenceResponse:
      properties:
        success:
          type: boolean
          title: Success
        created_files:
          items:
            type: string
          type: array
          title: Created Files
        skipped_files:
          items:
            type: string
          type: array
          title: Skipped Files
        message:
          type: string
          title: Message
          default: ''
      type: object
      required:
      - success
      title: InitPersistenceResponse
      description: Response from persistence initialization.
    ListBackupsRequest:
      properties:
        device_ip:
          type: string
          title: Device Ip
      type: object
      required:
      - device_ip
      title: ListBackupsRequest
      description: Request to list backups.
    ListBackupsResponse:
      properties:
        success:
          type: boolean
          title: Success
        config_backups:
          items:
            type: string
          type: array
          title: Config Backups
        hosts_backups:
          items:
            type: string
          type: array
          title: Hosts Backups
      type: object
      required:
      - success
      title: ListBackupsResponse
      description: Response with backup list.
...
    PortCheckRequest:
      properties:
        device_ip:
          type: string
          title: Device Ip
        timeout:
          type: number
          maximum: 60.0
          minimum: 1.0
          title: Timeout
          default: 10.0
      type: object
      required:
      - device_ip
      title: PortCheckRequest
      description: Request to check SSH port.
    PortCheckResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
        has_ssh:
          type: boolean
          title: Has Ssh
          default: false
      type: object
      required:
      - success
      - message
      title: PortCheckResponse
      description: Response with port check results.
...
    RestoreRequest:
      properties:
        device_ip:
          type: string
          title: Device Ip
        backup_path:
          type: string
          title: Backup Path
      type: object
      required:
      - device_ip
      - backup_path
      title: RestoreRequest
      description: Request to restore from backup.
    RestoreResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
      type: object
      required:
      - success
      - message
      title: RestoreResponse
      description: Response with restore result.
...
    WizardCompleteRequest:
      properties:
        device_id:
          type: string
          title: Device Id
          description: Device ID
      type: object
      required:
      - device_id
      title: WizardCompleteRequest
      description: Request to mark wizard setup as complete for a device.
    WizardCompleteResponse:
      properties:
        success:
          type: boolean
          title: Success
        device_id:
          type: string
          title: Device Id
        setup_status:
          type: string
          title: Setup Status
        message:
          type: string
          title: Message
      type: object
      required:
      - success
      - device_id
      - setup_status
      - message
      title: WizardCompleteResponse
      description: Response after marking wizard setup as complete.
...